requirements.txt
Flask==3.0.0
cryptography==41.0.7
pycryptodome==3.19.0
gmssl==3.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
